refactor: move service to provider/publiccloud
