+++ internal/provider/publiccloud/service/public_cloud/data_adapters/shared/value_adapters_test.go
