+++ internal/provider/publiccloud/service/public_cloud/data_adapters/to_data_source_model/adapter.go
@@ -3,8 +3,8 @@ package to_data_source_model
 import (
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 	"github.com/leaseweb/leaseweb-go-sdk/publicCloud"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/core/services/public_cloud/data_adapters/shared"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/data_sources/public_cloud/model"
+	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud/service/public_cloud/data_adapters/shared"
 )
 
 func AdaptInstances(sdkInstances []publicCloud.Instance) model.Instances {
