+++ internal/provider/publiccloud/service/public_cloud/data_adapters/to_data_source_model/adapter_test.go
