+++ internal/provider/publiccloud/service/public_cloud/data_adapters/to_opts/adapter_test.go
