+++ internal/provider/publiccloud/service/public_cloud/data_adapters/to_resource_model/adapter_test.go
