+++ internal/provider/shared/service/errors/error.go
