+++ internal/provider/shared/service/errors/error_test.go
