+++ internal/provider/shared/service/int_markdown_list.go
@@ -1,4 +1,4 @@
-package shared
+package service
 
 import (
 	"strconv"
